Change the event timestamp resolution to 64-bit
In SDL3 events have 64-bit timestamps. Preserve that resolution in
our events.

+++ src/events/InputEventGenerator.hh
@@ -51,7 +51,7 @@ public:
 private:
 	void handle(SDL_Event& evt);
 	void handleKeyDown(const SDL_KeyboardEvent& key, uint32_t unicode);
-	void splitText(uint32_t timestamp, const char* utf8);
+	void splitText(uint64_t timestamp, const char* utf8);
 	void setGrabInput(bool grab) const;
 
 	// EventListener
